+++ packages/core/src/index.ts
@@ -70,3 +70,4 @@ export * from "./timeline.ts";
 export * from "./opwords.ts";
 export * from "./canvassort.ts";
 export * from "./lens.ts";
+export * from "./export.ts";